[feat] 로그인 서버내용 받기
