+++ app/src/main/java/org/sopt/dosopttemplate/response/ResponseLoginDto.kt
@@ -1,0 +1,15 @@
+package org.sopt.dosopttemplate.response
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ResponseLoginDto(
+    @SerialName("id")
+    val id: Int,
+    @SerialName("username")
+    val username: String,
+    @SerialName("nickname")
+    val nickname: String
+)
+
